requirements.txt
boolean.py
nltk